Intial commit
This change adds a new package called `licensing` which provides a
set of functions that read the top lines of a file to determine if
they match a specific header.

introduces the `go-licenser` binary so that we can easily verify that
all the repository source files have a header with the license set.
The binary itself supports two modes of operation; a "dry run" mode
that is set with the boolean flag `-d` which will simply print which
files are missing the header to the standard output and the default
mode which simply rewrites the files to include the license header.
it detects if there's multiple licenses previously created licenses
don't match the current license format. It preserves any previous
package comments, build build flags or partial license headers.

Modifies the Makefile accordingly to include the `go-licenser -d`
check on `lint` and `go-licenser` in default mode on the `format`
target.

It also adds a new Makefile target `update-golden-files` that updates
the golden files / folder with `go test <package> -update`.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>