Add Error exit code decorator type
This change adds a new type Error that wraps any error and adds a code
property so that it can be read by the Code package function.
Also changes the signatures of all of the functions that returned int
and error and instead just returns the wrapped error type from which
the exitcode is obtained using the Code function.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>