Add -notice flag to autogenerate NOTICE files
Adds a -notice flag and a subset of flags tied to -notice. When the flag
is set, generates a `NOTICE` file at the root folder of the project when
used with or stdout when using `-d` for dry runs.

The NOTICE file generation heavily relies on the `go.mod` file and its
contents, Any other dependency managers outside of those reliant on
`go.mod` won't be able to make use of this feature.

By enabling this feature, the binary won't be zery dependency anymore
even though it's still very fast.

The format is fixed unless the Go template is overriden with the flag
`-notice-header`, in which case only the header is overwritten,
maintaining the list of dependencies and licenses within enclosed `=`.

It also changes the default error code from `0` to `255` any time the
error is not nil. Before any unspecified errorcodes defaulted to 0 which
is not the desired behaviour.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>