build: bump go version to 1.17 (#49)
* build: bump go version to 1.17

Remove deprecated GO111MODULE env variable and replace old go install
with the new 'go run module@version' syntax.
The new makefile is leaner and we can remove dependencies from go.mod
making the app truly 'zero dependency' as advertised.

* ci: remove old make deps step

* lint: fix staticcheck issues

Golint is deprecated and no longer maintained. Before replacing
it with staticcheck we should make sure the switch is not going
to create more warnings.
This commit should fix all staticcheck issues reported by the
latest version of the linter.

* refactor: replace deprecate ioutil calls

ioutil was deprecated in Go 1.16 and it is being reported
as a warning in staticcheck, blocking the update to Go 1.17.

* lint: run go fmt

* ci: run tests on the latest supported versions of go

CI was testing on ancient versions. Update the matrix to test
on the two supported versions (1.18 + 1.19) and also add 1.17
to be extra sure due to 1.19 being released recently.

* fix: update fixtures template and add support for new build tag lines

Go 1.17 added new build tag lines and they need to
be supported as end prefixes to avoid replacing them.

See https://go.dev/doc/go1.17#build-lines

* test: update fixture to fix failure